requirements.txt
Click==7.0
defusedxml==0.6.0
diff-match-patch==20181111
Django==2.2.6
django-ckeditor==5.8.0
django-crispy-forms==1.8.1
django-formtools==2.1
django-import-export==1.2.0
django-js-asset==1.2.2
django-pure-pagination==0.3.0
django-ranged-response==0.2.0
django-simple-captcha==0.5.12
et-xmlfile==1.0.1
Flask==1.1.1
future==0.18.2
httplib2==0.14.0
itsdangerous==1.1.0
jdcal==1.4.1
Jinja2==2.10.3
Markdown==3.1.1
MarkupPy==1.14
MarkupSafe==1.1.1
mysqlclient==1.4.6
odfpy==1.4.0
openpyxl==3.0.1
Pillow==6.2.1
PyMySQL==0.9.3
pytz==2019.3
PyYAML==5.1.2
six==1.13.0
sqlparse==0.3.0
tablib==0.14.0
Werkzeug==0.16.0
xlrd==1.2.0
xlwt==1.3.0
